confirmacion para eliminar (no anda)

+++ src/app/admin/admin.component.ts
@@ -1,7 +1,8 @@
+import { DialogsService } from './../shared/services/dialogs.service';
 import { Recorrido } from './../shared/models/recorrido';
 import { Colectivo } from './../shared/models/colectivo';
 import { AfService } from './../af.service';
-import { Component, OnInit } from '@angular/core'
+import { Component, OnInit, ViewContainerRef } from '@angular/core'
 import {Observable} from "rxjs/Rx";
 import {AngularFire, FirebaseListObservable} from 'angularfire2';
 
@@ -16,7 +17,11 @@ export class AdminComponent implements OnInit {
   colectivos: Colectivo[];
   recorridos: Recorrido[];
 
-  constructor(private afService: AfService) { 
+  constructor(
+    private afService: AfService,
+    private dialogsService: DialogsService, 
+    private viewContainerRef: ViewContainerRef
+    ) { 
     
   }
 
@@ -35,11 +40,28 @@ export class AdminComponent implements OnInit {
   }
 
   removeColectivo(id: string){
-    this.afService.removeColectivo(id);
+    var confirm = this.openDialog('Eliminar colectivo', '¿esta seguro de eliminar este colectivo?');
+    console.log('dialog result ', confirm)
+    if(confirm){
+      this.afService.removeColectivo(id);
+    }  
   }
 
   removeRecorrido(id: string){
-    this.afService.removeRecorrido(id);
+    if(this.openDialog('Eliminar recorrido', '¿esta seguro de eliminar este recorrido?')){
+      this.afService.removeRecorrido(id);
+    }  
+  }
+
+  openDialog(titulo, mensaje): boolean {
+    var result;
+    this.dialogsService
+      .confirm(titulo, mensaje, this.viewContainerRef)
+      .subscribe(function onNext(res){
+        result = res;
+        console.log(res);
+      });
+    return result;
   }
 
 }
